# project/openapi.yaml
openapi: 3.0.3
info:
  title: Globe Trotter API
  description: |
    Globe Trotter - Empowering Personalized Travel Planning
    
    Complete RESTful API for travel booking and itinerary planning platform.
    
    ## Authentication
    All protected endpoints require Firebase JWT token in Authorization header:
    `Authorization: Bearer <firebase-jwt-token>`
    
    ## Rate Limiting
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour
    - Premium: 5000 requests/hour
    
    ## Error Handling
    Standard HTTP status codes with consistent error response format.
  version: 1.0.0
  contact:
    name: Globe Trotter API Support
    email: api-support@globetrotter.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.globetrotter.com/v1
    description: Production server
  - url: https://staging-api.globetrotter.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server
security:
  - FirebaseAuth: []
paths:
...
  /destinations:
    get:
      tags: [Destinations]
      summary: Search destinations
      operationId: searchDestinations
      security: []
      parameters:
        - name: query
          in: query
          description: Search query (name, city, country)
          schema:
            type: string
        - name: latitude
          in: query
          description: Latitude for location-based search
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          description: Longitude for location-based search
          schema:
            type: number
            format: double
        - name: radius
          in: query
          description: Search radius in kilometers
          schema:
            type: integer
            default: 50
        - name: country
          in: query
          description: Filter by country
          schema:
            type: string
        - name: min_rating
          in: query
          description: Minimum rating filter
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: categories
          in: query
          description: Activity categories (comma-separated)
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        200:
          description: Destinations found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationSearchResponse'
        400:
          $ref: '#/components/responses/BadRequest'
  /destinations/{destinationId}:
    get:
      tags: [Destinations]
      summary: Get destination details
      operationId: getDestination
      security: []
      parameters:
        - name: destinationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Destination details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DestinationResponse'
        404:
          $ref: '#/components/responses/NotFound'
...
  /trips:
    get:
      tags: [Trips]
      summary: Get user trips
      operationId: getUserTrips
      parameters:
        - name: status
          in: query
          description: Filter by trip status
          schema:
            type: string
            enum: [draft, planned, active, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: User trips retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListResponse'
    post:
      tags: [Trips]
      summary: Create new trip
      operationId: createTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreateRequest'
      responses:
        201:
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        400:
          $ref: '#/components/responses/BadRequest'
...
  /bookings:
    get:
      tags: [Bookings]
      summary: Get user bookings
      operationId: getUserBookings
      parameters:
        - name: status
          in: query
          description: Filter by booking status
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: type
          in: query
          description: Filter by booking type
          schema:
            type: string
            enum: [flight, hotel, activity, transport, package]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        200:
          description: User bookings retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
    post:
      tags: [Bookings]
      summary: Create booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        201:
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        400:
          $ref: '#/components/responses/BadRequest'
  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get booking details
      operationId: getBooking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        404:
          $ref: '#/components/responses/NotFound'
  /bookings/{bookingId}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel booking
      operationId: cancelBooking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Cancellation reason
      responses:
        200:
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        404:
          $ref: '#/components/responses/NotFound'

  # Recommendations
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase JWT token
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages

    # User schemas
...
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        google_place_id:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        country:
          type: string
        city:
          type: string
        continent:
          type: string
        description:
          type: string
        short_description:
          type: string
        avg_rating:
          type: number
          minimum: 0
          maximum: 5
        review_count:
          type: integer
        average_price:
          type: number
        cost_of_living_index:
          type: integer
        safety_index:
          type: integer
        avg_temperature:
          type: number
        activity_categories:
          type: array
          items:
            type: string
        featured:
          type: boolean
    DestinationSearchResponse:
      type: object
      properties:
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/DestinationResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Trip schemas
    TripCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        traveler_count:
          type: integer
          minimum: 1
        total_budget:
          type: number
        currency:
          type: string
          minLength: 3
          maxLength: 3
        privacy_level:
          type: string
          enum: [private, friends, public]
...
    TripResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        traveler_count:
          type: integer
        status:
          type: string
        total_budget:
          type: number
        currency:
          type: string
        privacy_level:
          type: string
        stops:
          type: array
          items:
            $ref: '#/components/schemas/TripStopResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TripListResponse:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: '#/components/schemas/TripResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    TripStopResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        destination_id:
          type: string
          format: uuid
        destination:
          $ref: '#/components/schemas/DestinationResponse'
        sequence_order:
          type: integer
        arrival_date:
          type: string
          format: date
        departure_date:
          type: string
          format: date
        notes:
          type: string
        budget:
          type: number

    # Booking schemas
    BookingCreateRequest:
      type: object
      required: [booking_type, amount, booking_details]
      properties:
        trip_id:
          type: string
          format: uuid
        booking_type:
          type: string
          enum: [flight, hotel, activity, transport, package]
        amount:
          type: number
        currency:
          type: string
        booking_details:
          type: object
        booking_date:
          type: string
          format: date
        service_date:
          type: string
          format: date
    BookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trip_id:
          type: string
          format: uuid
        booking_type:
          type: string
        external_booking_id:
          type: string
        status:
          type: string
        amount:
          type: number
        currency:
          type: string
        payment_status:
          type: string
        payment_provider:
          type: string
        booking_details:
          type: object
        booking_date:
          type: string
          format: date
        service_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
    BookingListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # Review schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "Invalid request parameters"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
...